requirements.txt
fastapi==0.104.1
google-generativeai
pydantic
mangum
uvicorn